# Speedup cloud reads