Re-run benchmarks after #317 is merged