@hombit 2025-07-27
Benchmarks copy-pasted from the action


```
| Change   | Before [8decb32f]    | After [0f459787]    |   Ratio | Benchmark (Parameter)                               |
|----------|----------------------|---------------------|---------|-----------------------------------------------------|
|          | 913M                 | 992M                |    1.09 | benchmarks.ReadFewColumnsS3.peakmem_run             |
|          | 10.2±0.07ms          | 10.2±0.05ms         |    1.01 | benchmarks.NestedFrameQuery.time_run                |
|          | 250M                 | 250M                |    1    | benchmarks.AssignSingleDfToNestedSeries.peakmem_run |
|          | 135M                 | 135M                |    1    | benchmarks.CountNestedBy.peakmem_run                |
|          | 101M                 | 102M                |    1    | benchmarks.NestedFrameAddNested.peakmem_run         |
|          | 10.6±0.1ms           | 10.6±0.2ms          |    1    | benchmarks.NestedFrameAddNested.time_run            |
|          | 106M                 | 106M                |    1    | benchmarks.NestedFrameQuery.peakmem_run             |
|          | 105M                 | 105M                |    1    | benchmarks.NestedFrameReduce.peakmem_run            |
|          | 270M                 | 270M                |    1    | benchmarks.ReassignHalfOfNestedSeries.peakmem_run   |
|          | 66.3±0.7ms           | 64.9±0.2ms          |    0.98 | benchmarks.CountNestedBy.time_run                   |
|          | 43.3±0.9ms           | 42.4±0.5ms          |    0.98 | benchmarks.ReassignHalfOfNestedSeries.time_run      |
|          | 1.40±0.01ms          | 1.35±0.02ms         |    0.96 | benchmarks.NestedFrameReduce.time_run               |
|          | 32.3±0.6ms           | 29.8±0.5ms          |    0.92 | benchmarks.AssignSingleDfToNestedSeries.time_run    |
| -        | 212M                 | 177M                |    0.84 | benchmarks.ReadFewColumnsHTTPS.peakmem_run          |
| -        | 3.61±0.04s           | 2.74±0.03s          |    0.76 | benchmarks.ReadFewColumnsHTTPS.time_run             |
| -        | 1.79±0.01s           | 741±20ms            |    0.41 | benchmarks.ReadFewColumnsS3.time_run                |
```